GoRackHD Makefile 2.0
This patch improves the Makefile so that a two-phase approach is used
for downloading project dependencies. The first phase uses the file
`glide-swagger.yaml` to download *only* the requested version of
go-swagger that is necessary to generate the client bindings. Once the
bindings are generated the second phase uses the standard `glide.yaml`
to scan the entire project to fetch the remaining dependencies, this
time including the `client` and `models` packages.