Return to using vendored go-swagger
The Makefile is using go-swagger out of $GOPATH instead of out the
vendor/ directory. gorackhd is supposed to use everything out
of vendor/, so there's no reason to reference a GOPATH'd version
of go-swagger in the Makefile at all.

We had unknowingly been using newer versions of go-swagger to generate
client code, so this builds on top of that by just setting the version
to 0.5.9, which is pre-release of 0.6.0. Alternative would be to set
version to 0.5.0 and go with that, but let's move on to the future!

We also start checking in glide.lock so our deps are locked. The
behavior to date has been that even when locking down go-swagger to
0.5.9, regenerating the client lib has produced different files at
times -- presumably because of changing deps. This will allow us to
have truly versioned releases of gorackhd. Since we are using
glide.lock, we don't need to take a two-phased appraoch anymore. We just
install all the dependencies with "glide install", which is faster as
well since no dependency resolution is done. This means a bit more work
on the development sinde when we want to update deps (meaning that it
can't all be done through the Makefile), but I think that is a useful
tradeoff.

The test target has also been simplified by removing the cover
pieces. The existing code through errors about not producing a cover
profile on mulitiple packages. This functionality is not currently
being used, so remove it for now.