Move to using go-swagger 0.6.0
We can now standardize on go-swagger 0.6.0. This version of go-swagger
changes things a a bit, however, in that it introduces a timeout to all
API calls. In order to have the timeout initialized by default, we need
to make sure we use the helper methods to initialize request params,
rather than initializing them directly.